+++ HISTORY.md
@@ -2,6 +2,8 @@
 
 # not yet published, version 9.4.6
 
+- Fix #2319: make the API of `Parser.evaluate` consistent with `math.evaluate`: 
+  support a list with expressions as input.
 - Improved documentation of function `setCartesian`. Thanks @fieldfoxWim.